mudança de L para int

+++ src/entity/Adm.java
@@ -1,7 +1,7 @@
 package entity;
 
 public class Adm extends Usuario{
-    public Adm(Long codigo, String nome, String email, String senha, String tipo) {
+    public Adm(int codigo, String nome, String email, String senha, String tipo) {
         super(codigo, nome, email, senha, tipo);
     }
 }
